requirements.txt
streamlit>=1.31.0

langchain>=0.1.0
langchain-community>=0.0.24
langchain-openai>=0.0.6

chromadb>=0.5.0
pypdf>=4.0.0

python-dotenv>=1.0.0
tiktoken>=0.5.2